% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_v_total_plot.R
\name{baseVtotal_plot}
\alias{baseVtotal_plot}
\title{Percent contribution of base peak intensity to total ion intensity for all MS2 events}
\usage{
baseVtotal_plot(mzXMLHeader)
}
\arguments{
\item{mzXMLHeader}{dataframe returned by mzR::header()}
}
\value{
plotly plot
}
\description{
Percent contribution of base peak intensity to total ion intensity for all MS2 events
}
