% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{createReport}
\alias{createReport}
\title{Title}
\usage{
createReport(readPath, writePath)
}
\arguments{
\item{readPath}{path to mzML/mzXML}

\item{writePath}{path and file name of whereto write html output}
}
\value{
report
}
\description{
Title
}
