% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_tic_max_bbpc.R
\name{getBpc}
\alias{getBpc}
\title{Get the Base Peak Chromatogram intensities for a given polarity and MS level}
\usage{
getBpc(mzHeader)
}
\arguments{
\item{mzHeader}{mzR header (data.frame)}
}
\value{
numeric vector of intensities
}
\description{
Get the Base Peak Chromatogram intensities for a given polarity and MS level
}
