% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_tic_max_bbpc.R
\name{getTic}
\alias{getTic}
\title{Get the total ion chromatogram  intensities for a given polarity and MS level}
\usage{
getTic(mzHeader)
}
\arguments{
\item{mzHeader}{mzR header (data.frame)}
}
\value{
numeric vector of intensities
}
\description{
Get the total ion chromatogram  intensities for a given polarity and MS level
}
