% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepForPlotly.R
\name{prepForPlotly}
\alias{prepForPlotly}
\title{Prepare mzR header data for plotly}
\usage{
prepForPlotly(mzPointer, msLevel, polarity)
}
\arguments{
\item{mzPointer}{mzPointer}

\item{msLevel}{msLevel}

\item{polarity}{polarity (in mzR: 0= negative, 1=positive, -1=unkown)}
}
\description{
Prepare mzR header data for plotly
}
