% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mz.R
\name{readMzSpec}
\alias{readMzSpec}
\title{Get mzR header data from selected mzXML file}
\usage{
readMzSpec(mzPointer)
}
\arguments{
\item{mzPointer}{mzR object}
}
\value{
dataframe
}
\description{
Get mzR header data from selected mzXML file
}
