% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_relative.R
\name{relativateTotalAndBase}
\alias{relativateTotalAndBase}
\title{For y-axis scaling}
\usage{
relativateTotalAndBase(maxTic, maxBpc)
}
\arguments{
\item{maxTic}{maximum in tic intensity vector}

\item{maxBpc}{maximum in bpc intensity vector}
}
\value{
numeric of length one
}
\description{
For y-axis scaling
}
