% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_vs_base_mz_plot.R
\name{rtVsBaseMzPlot}
\alias{rtVsBaseMzPlot}
\title{Retention time vs precursorMZ}
\usage{
rtVsBaseMzPlot(input)
}
\arguments{
\item{input}{dataframe}
}
\value{
ggplot
}
\description{
Retention time vs precursorMZ
}
