% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_vs_prec_mz_plot.R
\name{rtVsPrecMzPlot}
\alias{rtVsPrecMzPlot}
\title{Retention time vs precursorMZ}
\usage{
rtVsPrecMzPlot(input)
}
\arguments{
\item{input}{dataframe}
}
\value{
ggplot
}
\description{
Retention time vs precursorMZ
}
