% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic_bpc_plot.R
\name{ticBpcPlot}
\alias{ticBpcPlot}
\title{TIC and BPC plot}
\usage{
ticBpcPlot(input, relInt, mtic, mbpi)
}
\arguments{
\item{input}{dataframe}

\item{relInt}{relative intensity correction}

\item{mtic}{for scaling y axes}

\item{mbpi}{for scaling y axes}
}
\value{
ggplot
}
\description{
TIC and BPC plot
}
